Add CI-Fuzz project configuration and first fuzz targets
The project config file .code-intelligence/project.yaml is read by the
CI-Fuzz fuzzing server and contains information about the sanitizers and
fuzz engines to be used, the runtime for the fuzz targets, the docker
image to use and some additional options. For more information see the
comments in the file.

The script .code-intelligence/build.sh tells the CI-Fuzz server how to
build the project.

The folder .code-intelligence/fuzz-targets contains some simple fuzz
targets. The *.yaml files specify which build flags to use for building
the fuzz target.